test(encryption tab): move e2e test which are testing all the encryption tab in `encryption-tab.spec.ts
