+++ playwright/snapshots/settings/encryption-user-tab/encryption-tab.spec.ts/default-tab-linux.png
